MB-28453: Switch xattrs from byte_buffer to char_buffer
KV-engine is unifying xattrs to use char_buffer, so dbdump.cc needs
a small change to remove an now unnecessary cast.

From the related kv-engine change:
<snip>
Overall we wish to remove the use of byte_buffer as we frequently
end up casting it to char. This commit just switches xattr code from
using a byte_buffer to char_buffer, making it easier to add
decompression support to xattr::Blob which is required to cleanly
address the MB this commit is linked with.
</snip>

Change-Id: I197723539ea1f7c799ccf3dcddfa1757531e36ba
Reviewed-on: http://review.couchbase.org/90422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>