UBSan: disable vptr check for V8 code
Our build of V8 doesn't include RTTI information. This means that
UBSan's vptr checks fail; as they cannot determine the type of
pointers created by V8. For example:

    mapreduce.cc runtime error: member call on address 0x7ffff1e24060 which does not point to an object of type 'Allocator'
    0x7ffff1e24060: note: object has invalid vptr

To allow us to use UBSan in the rest of couchstore; add a macro to
disable the specfic VPTR checks; and annotate functions which exhibit
the problem.

Change-Id: I85df513847b48d7d638e71627f0837305fc83302
Reviewed-on: http://review.couchbase.org/95820
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>