MB-31838: Couchstore- Log messages when a corruption is detected
This addresses the MAC build failure due to use of thread_local.
On master MB-31767: Couchstore- Log messages when a corruption is detected

When a corruption is detected, log the file offset, the checksum
and length fields. When an index node corruption was seen it
showed checksum and length fied corruption. Some corruptions
indicated the checksum was ok but snappy uncompression failed
to uncompress. If the length and checksum are 0, it is possible
for checksum validation to pass, but the snappy uncompress will
report an error. The new log will help verify if the length is 0.

When writing, snappy compression failures and unexpected increase
in size after compression are logged. The size increase is possible
but rare.

A new interface "couchstore_last_internal_error" which reports the last
internal error in case of a failure is returned.

There is a test case to inject an error to corrupt the length and
checksum to 0. This will not result in a checksum error but a
corruption error.

Old Change-Id: I2c3788510291b6eee7f43b1b22f0ed6eabe2359c

Change-Id: Ic4c11531bd9b2f7f4c9c3b234c8b306555d0529b
Reviewed-on: http://review.couchbase.org/101447
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>