MB-34862: couchstore- return of 0 by pwrite() is not handled
A return value of 0 from pwrite() would actually mean that
no bytes were written and the caller should retry pwrite.
Couchstore code does not retry and it leads to assuming an
entire block has been written out when it hasn't.

The fix is to retry in the case of 0 bytes getting written.

Unit tests for buffered and unbuffered cases are added.

Change-Id: I5f592d8a4084193e6bc9df88a1220a84854d113e
Reviewed-on: http://review.couchbase.org/112505
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>