MB-35473: couchstore- return of 0 by pwrite() is not handled
Backport of MB-34862

A return value of 0 from pwrite() would actually mean that
no bytes were written and the caller should retry pwrite.
Couchstore code does not retry and it leads to assuming an
entire block has been written out when it hasn't.

The fix is to retry in the case of 0 bytes getting written.

Unit tests for buffered and unbuffered cases are added.

Change-Id: I71a1e8a247edfeb773db93906a70a6624de252b9
Reviewed-on: http://review.couchbase.org/113353
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>