MB-36424: Ensure flush_mr_partial() obeys node size quota
When couchstore writes out modified node elements during document
saves, it it supposed to limit the number of bytes written in a single
node to a _chunk_threshold - by default 1279 bytes. If the node is
larger than the limit it should be split into multiple sibling nodes.

However, this limit is _not_ respected, resulting in overly-large
nodes being written out. In the case of the by-seqno B-Tree (which
always writes values to the rightmost leaf as seqnos are increasing),
it results in all leaf elements residing in a single leaf
node. Moreover, this means that adding another element to the B-Tree
effectively re-writes the entire tree, resulting in massive Write
Amplification. Local testing saw Write Amplification up to 743x,
compared to around 5x immediately after compaction (when B-Tree
structure is repaired).

Problem is in the handling of the node quota in flush_mr_partial() -
the while loop terminating conditions is if mr_quota goes below zero
(i.e. quota has been consumed). mr_quota is provided as an input
parameter, and then decremented inside the while loop.

However, the type of mr_quota is an unsigned size_t - as such when it
is decremented in the while loop after writing the key/value pair it
goes to a number close to SIZE_T_MAX, not a negative value. As a
result the while loop is not exited.

This bug has been present since July 2012 - was introduced via
http://review.couchbase.org/#/c/18008/

Change-Id: I60200fcb100525ceb0fd91971c702645f0c4f3f5
Reviewed-on: http://review.couchbase.org/116219
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>