MB-39629: Move fetch_callback_ctx to couchfile_modify_request
Currently we use the arg field of couchfile_modify_action for a
context object to pass into the fetch_callback. As this is only used
in one place and every context object is identical, refactor the
couchfile_modify_request to instead take a pointer to the
callback_ctx and remove the now unnecessary union from
couchfile_modify_action. This will allow us to combine Fetch and Insert
actions in a future patch which will save us memory overhead when
updating indexes as part of flushing.

Change-Id: If1c561ed3f3c517fb97f0cba2fbc02cfc28ac4be
Reviewed-on: http://review.couchbase.org/120030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>