MB-25982: Reduce memory usage during compaction
Compaction creates the ID index using merge sort of the keys.
To create the sorted runs a certain number of keys are loaded
into memory and sorted. The value passed in as the number of
keys to load is large. The fix is to base the number of keys
on the memory to use and the average key size for the ID index.
The assumption is the average key size is 200bytes (key,value
and overhead) and 100MB of memory available for sort.

Change-Id: If9a750a08659207c5e47235878b30314e96b793d
Reviewed-on: http://review.couchbase.org/120601
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>