Refactor: change seek to work with UniqueDbPtr
The old seek methods would hide the original error message
and just close (and release) the database handle and give
a generic error message back. This made them "weird" to use
with unique-pointers as they would be "closed and freed"
and you would have a "magic" call to release as the object
was already destroyed.

The new seek will try to reload the header it used to have
open upon failures. If that fails for some reason the
database is closed (but not freed) and the error
COUCHSTORE_ERROR_DB_NO_LONGER_VALID is returned. In all
other error situations you'll get the actual error (and
the database handle may still be used).

Given that we don't _free_ the database instance as part
of returning COUCHSTORE_ERROR_DB_NO_LONGER_VALID it plays
nicely with a UniqueDbPtr as you may call
couchstore_close_db() on a closed database.

Change-Id: Iffe7b22f617dfb577af112be04a98f34d505c1de
Reviewed-on: http://review.couchbase.org/c/couchstore/+/124750
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>