Add support for PiTR to compaction
A typical compaction would rewrite a database file looking like:

    |H|D|D|D|H|D|H|D|D|D|H|D|D|H|

to something like

    |D|D|D|D|D|D|D|D|D|H|

File format 13 of CouchStore contains a timestamp in the header
block which identifies when the data was written (the application
may provide this timestamp as part of commit). When compacting
with support for Point in Time Recovery (PiTR) compaction won't
remove all of the intermediate header blocks in the file, but
do a full compaction up to the oldest time we want to keep around,
then perform incremental compaction from that point forward keeping
only a subset of the headers around.

If we assume a file with the following timestamps in its headers:

    0, 10, 12, 14, 15, 19, 21, 31, 40

If we want to compact with the oldest timestamp set to 11 and the
interval to 10 the following would happen:

    * Full compaction up to the header with timestamp 10. There
      is no header with timestamp 11, so we need to preserve the
      one before that.

    * squash header 12, 14, 15 and 19 together into a new
      header with timestamp 19

    * preserve header 21, 31 and 40 as they can't be squashed

The final database would then look like:

    10, 19, 21, 31, 40

Change-Id: I598b62aa0cd704e17e7312f6c05414d48a0c6646
Reviewed-on: http://review.couchbase.org/c/couchstore/+/125121
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>