MB-37319: Add new method to seek for a seqno
seekFirstHeaderContaining allows a caller to seek for the first
header which contains the provided sequence number. To make it
more useful for people who use granularity one may also specify
a granularity to allow the method to align that header to the
provided granularity (to allow the client to do some
deduplication). The granularity is aligned into barriers with
offsets relative to "epoch" and calculated from:

   low = timestamp - timestamp % granularity
   high = low + granularity

And the following range is used [low,high>

Change-Id: I7689c54e98d8b75683b44bc9088ecf7b2c25d3ee
Reviewed-on: http://review.couchbase.org/c/couchstore/+/127528
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>