Cleanup: Reduce the scope for variables
couchstore_save_documents_and_callback was initially written
in C (and using goto) so it needed all variables to be declared
before goto. Now that we're using C++ we can delay that until
the variables are used (and use early return). Switch to a
unique_ptr for the fatbuff to make sure we don't leak memory

Change-Id: I07f2687379b97f6f9fbb66149639ac7f4ea20ebe
Reviewed-on: http://review.couchbase.org/c/couchstore/+/127714
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>