MB-36335: Add saveLocalDocuments
Making many calls to couchstore_save_local_document in one
commit batch can lead to poor utilisation of the file blocks
in database file as each call writes a new index. With collections
support in kv-engine, use of the local index has increased possibly
with 1K local doc writes in a single batch.

By adding a new function that allows for many local documents
to be written in one index update we achieve much better utilisation
in the database file.

Change-Id: I106caa557ac2cfda789b2734c59ed795559f9c1f
Reviewed-on: http://review.couchbase.org/c/couchstore/+/130057
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>