MB-40030: Fix scope/collection event key parsing
Currently we don't parse system event keys correctly, and we end up just
print "id: (system-event-key:) " for all system event keys. We should
instead parse them correctly to work out if their for a scope or
collection and extract the ID correctly. Thus, printing them in the
following format: "id: (system-event-key:collection:0xa) _collection".

To do this we need to parse the leb128 parts of the key correctly as
they are laid out in the following format:
[leb128 system-collection-cid][leb128 bool 1 for scopeID. 0 for
collection][leb128 for cid/sid][_collection or _scope].

Change-Id: I8bee67f968c7d5bb84316cb9649f91d5d25a49a7
Reviewed-on: http://review.couchbase.org/c/couchstore/+/133622
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>