Allow seek(Forward) to pick up additional writes
Allow a read-only instance of the database to seek to headers added
by other writers after the file was opened.

This is an optimization needed by running compaction in parallel with
the flusher to avoid having to close and reopen the file just to
check if there is more data added to the file.

Change-Id: Ided24016eb1d4846369cb4470cba068eb8c8dc2e
Reviewed-on: http://review.couchbase.org/c/couchstore/+/136345
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>