Add seek(END) to allow "resize" of the file
As part of running compaction in parallel with the flusher
we want to get the "latest" header of the database in the
catch up phase.

It is more efficient to use a seek(End) than trying to
go header by header forward in the file (typically the
header is the last block in the file so we wouldn't have
to check each 4k block from the current offset until we hit
the tail of the file).

Remove the "implicit" resize as part of Forward as it isn't
needed in the current code and it adds complexity to the
code.

Change-Id: I515a513f6606d1b4bb8d21fa41f69f5e3c41a987
Reviewed-on: http://review.couchbase.org/c/couchstore/+/138067
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>