MB-39946: Change how physical_size is calculated
The old calculation which used new - old positions meant that
physical_size could sometimes include the 1 byte "block prefix".

This makes it harder to use in the context of collections (now
the major user of physical_size).

Change the calculation so that the same document always results
in the same physical_size regardless of block boundaries.

Change-Id: Ic7661e0c05dfcac236d44da05e6bad0e23ec355c
Reviewed-on: http://review.couchbase.org/c/couchstore/+/141085
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>