MB-42224: Implememt re-try loop for header's fsync failure
A data-write in couchstore is a 4-step operation:
  1) pwrite data to OS cache
  2) fsync data to disk
  3) pwrite header to OS cache
  4) fsync header to disk

Before MB-42224, any failure at any step was handled by returning an
error code to EP. EP was responsible for re-trying the call.
That approach is perfectly fine if we hit a failure at step (1), (2) or
(3). The new data is even fsync'ed to disk, but it's not visible as
there is no header that indexes the new data (as step (3) has never
succeded).

Problem is if (1)->(3) succeed and we fail at (4). Both new data and
header have been written to the OS cache, so new data is visible when
EP re-tries the call.
That causes a number of side effects on stats computation. Eg when EP
re-tries an insert, couchstore sees data from the failed commit and
marks the insert as an update, so we miss to increment the num-items.

Our solution is to re-try the header flush if step (4) fails.

Change-Id: I4554801cff956f5cc32392705ba51cbc881581a8
Reviewed-on: http://review.couchbase.org/c/couchstore/+/145098
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>