MB-42224: Don't use folly::fdatasyncNoInt in PosixFileOps::sync
Call introduced in http://review.couchbase.org/c/couchstore/+/144932,
that is responsible for a perf degradation in EP unit tests.
Tracing shows that the functions introduces syscalls to fcntl, which is
hot in the measurement.

Fix is just to make a direct syscall to fdatasync.

Note that docs of fsync/fdatasync don't report EINTR as possible errno,
so it seems that we don't really need any specific re-try mechanism.
Anyway, EINTR will be still handled in the main error code path.

Change-Id: Ia0ea405b65a51132ba584c5e74359ff14bfc4264
Reviewed-on: http://review.couchbase.org/c/couchstore/+/146096
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>