MB-41257: Make WindowsFileOps::open() use oflags to decide R/W mode
Currently WindowsFileOps::open() ignores the oflags state when calling
CreateFileA(), and instead sets dwDesiredAccess to always be in
read-write mode (GENERIC_READ | GENERIC_WRITE). This is incorrect as
couchstore sometimes may want to open the file on-disk in read or write
only mode. To fix this check the ofags to decide that the value of
dwDesiredAccess should be.

Change-Id: Ic177b6ee83a2fd5d7fef8dc3b92e5f2114955948
Reviewed-on: http://review.couchbase.org/c/couchstore/+/149724
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>