Move to_json(Header) to own file
<libcouchstore/couch_db.h> costs on average 0.5s to include and is
included 31 times. Of the 0.5s to include; the majority of the time is
spent including <nlohmann/json_fwd.hpp> (and it's
dependancies). However, nlohmann::json is used just once in
couch_db.h; for a utility function to convert Headet to JSON. That
utility function is used just once in dbdump.cc.

Move to_json into its own header file (json_utils.h) which the callers
which do need it can include. This reduces the average include cost of
couch_db.h significantly.

Change-Id: Idf177e1c97f62223766348c55bee0dea33f80ab3
Reviewed-on: http://review.couchbase.org/c/couchstore/+/156481
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>