MB-49817: Ensure --json outputs valid json
Ensure that when --json arg is used couch_dbdump outputs valid json. To
ensure this happens move our rending of the json from multiple printf
calls, to instead use nlohmann::json.

This also makes a key change to make the json we output valid to store
the key:'content_meta' value as a string so it can still be rendered as
a hex value.

Change-Id: I901d3deac092a8590b0cfeeac78e9fe61d673dfe
Reviewed-on: https://review.couchbase.org/c/couchstore/+/167549
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>