MB-51369: Header0 timestamp is 0 on empty data file
Couchstore data files might be created "empty" at open. If that's the
case, then an initial header is created at pos=0.

Currently the timestamp for that header is set to the real timestamp at
file creation. That is currenlty breaking the timestamp monotonicity
invariant in theoretically all the test that encode the commit
timestamp manually. Also, that is making new tests hard to code, eg the
new PiT tests in https://review.couchbase.org/c/couchstore/+/172959.

This patch changes our logic so that the header at pos=0 has
timestamp=0. Note that Header0 is a placeholder that logically points to
no data.

Change-Id: I185a073087e8e2bbfd2d1243c9fe07868fa7ad21
Reviewed-on: https://review.couchbase.org/c/couchstore/+/173641
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>