MB-53923: couch_dbdump: Output CAS as string in JSON mode
couch_dbdump --json outputs CAS and other fields which are internally
64bit integers as JSON number types.

While the JSON library we use (nlohmann::json) has no problem with
64bit integers, other JSON libraries / tools (such as jq) do not
correctly parse these large numbers and silently discard
least-significant bits / round the value. For example, note the
difference in displayed CAS value with and without piping the output
of couch_dbdump to jq :

    install/bin/couch_dbdump --key  doc_key --json ../ns_server/data/n_0/data/default/997.couch.1
    {"body":{"json":"maybe?"},"cas":1664543368459911168, ...
vs

    install/bin/couch_dbdump --key doc_key --json ../ns_server/data/n_0/data/default/997.couch.1 | jq -c
    {"body":{"json":"maybe?"},"cas":1664543368459911200, ...

The correct CAS value of 1664543368459911168 is mangled and is rounded
to 1664543368459911200.

In the specific case of jq this is a known bug and fixed in master,
but is not fixed in the current latest release (1.6). To try to
mitigate these kinds of problems in future, change couch_dbdump to
represent large numeric values as stings which will not be mangled.

Change-Id: I83479399fe9eb70de08644741d45794a7681f87e
Reviewed-on: https://review.couchbase.org/c/couchstore/+/180662
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker