MB-54297: Fix heap mismatch when libcouchstore.so loaded as python library
+Problem+

Since rebuilding Folly to use GCC-10.2 via tlm patch: "MB-54162:
cbdeps: Build folly with same compiler as Server" (63cfd75),
cbworkloaden has started to crash during loading libcouchstore.so via
the couchstore Python library.

The crux of the issue is that GCC-10.2 makes different inlining
decisions for the ISRA (interprocedural scalar replacement of
aggregates) optimisation pass, which results in functions related to
std::string construction (_M_create) which were previously inlined
into libfolly (and consequently libcouchstore.so) are no longer
inlined, and the version from libstdc++.so.6 is used instead.

This is significant because folly's static initialiser code (executed
when libcouchstore.so is loaded) creates some std::string objects,
which in turn call a helper std::string::_M_create() function which in
turn calls 'operator new' to allocate heap memory for the constructed
string.

- Previously, as _M_create() was inlined into libcouchstore.so, the
  'operator new' symbol was found inside libcouchstore.so - i.e. our
  version from global_new_replacement.cc which redirects allocations
  to jemalloc:

    libcouchstore.so: follyCacheLocality::readFromProcCpuinfo
    libcouchstore.so:  -> basic_string::basic_string() [inlined]
    ...
    libcouchstore.so:     -> basic_string::_M_create() [inlined]
    libcouchstore.so:	     -> operator_new
    libjemalloc.so:             -> je_malloc

- With GCC-10.2, as _M_create() is not inlined and is found within
  libstdc++.so.6, lookup of 'operator new' symbol is done from the
  scope of libstdc++.so. If libstdc++.so has already been loaded
  before libcouchstore.so (by another C++ python library such as
  snappy), then the operator new symbol in libstdc++.so will have
  already been bound to the first definition available - i.e. the one
  in libstdc++.so - which allocates memory via malloc():

    libcouchstore.so: follyCacheLocality::readFromProcCpuinfo
    libcouchstore.so:  -> basic_string::basic_string() [inlined]
    ...
    libstdc++.so.6:       -> basic_string::_M_create()
    libstdc++.so.6:	     -> operator_new
    libc.so:                    -> malloc

As a result, when we later free this string from libcouchstore.so and
hence via the custom operator new which uses jemalloc, there's a
mismatch when jemalloc attempts to free memory which it did not
allocate, and we segfault.

+Analysis+

This problem arises due to incorrect assumptions about how memory will
be allocated from libcouchstore.so when it is loaded as a plugin by
programs which are using different definitions of operator new /
delete - such as the python interpreter.

We had assumed that as long as libcouchstore.so overrides operator new
/ delete itself; any allocations (or deallocations) it makes will go
via the overridden functions. This assumption is clearly incorrect -
if some other library (e.g. libstdc++.so) allocates memory "on behalf
of" libcouchstore.so, and that library already has bound operator new,
then the existing definition of operator new will be used.

It's worth noting that no issues were seen with couchstore itself or
it's unit tests; as while we /do/ dlopen() libcouchstore.so from
python, there is no other libraries using libstdc++.so already
imported and so the initial bind of 'operator new' inside libstdc++.so
is done with the override in libcouchstore.so in scope and that one is
selected.

Ultimately our assumptions about operator new/delete overriding are
blown, we cannot safely use overriden operator new / delete from
libcouchstore.so if it is going to be loaded from python (or any other
similar environment).

+Why does libcouchstore.so need to override new/delete?+

Recall that we override operator new/delete for two reasons:

1. To be able to track the amount of heap memory in use on a
   per-Bucket basis for kv_engine - override all calls to operator
   new/delete we can add our own accounting.

2. To replace system heap with a higher performance implementation
   (jemalloc) for performance sensitive parts of the system.

Given that couchstore is used by KV-Engine and (1) makes memory
allocations we want to track and (2) benefits from a higher
performance malloc impl, we historically overrode new/delete inside
libcouchstore.so.

However, since it was initially added, we have moved to linking
kv_engine with couchstore statically, using the couchstore_static.a
static library. This was done due to MB-45378 (issues with multiple
instances of libfolly in the process) and we needed to convert all
consumers of Folly (such as couchstore) to be statically linked.

This is somewhat fortuitous as it means that kv_engine no longer links
to libcouchstore.so, and hence requirement 1 (memory tracking) is no
longer applicable to libcouchstore.so. Regardigng requirement 2
(higher performance heap), libcouchstore.so is used by te following
programs: higher-performance malloc:

    $ ninja -t query couchstore/libcouchstore.so | grep -A 100 'outputs:'
      outputs:
	couchstore/couch_create
	couchstore/couchstore_gtest
	couchstore/couch_dbinfo
	couchstore/couch_compact
	couchstore/couch_dbck
	couchstore/programs/pitr_compact/pitr_compact
	kv_engine/couchfile_upgrade

None of these programs _need_ a faster heap, and so there shouldn't be
any issue in switching from jemalloc back to the system heap for them.

Note that until recently libcouchstore_views.so (the map/reduce
variant of couchstore) also used global_new_replacement.cc to use
jemalloc by default, for performance reasons. However, as of MB-54106
(v7.1.2) it no longer uses global_new_replacement.cc due to similar
heap mismatch issues inside V8. There was no performance impact
observed when switching back to system heap from jemalloc, which gives
confidence that the above couchstore tools should not be significantly
impacted when the same is done to them.

+Solution+

Modify libcouchstore.so to no longer override operator new/delete, by
linking to platform_without_global_new_replacement instead of
platform). This is the same solution which the view engine took for
the aforementioned MB-54106 V8 heap mismatch issue.

Change-Id: I71513c8c221f13767b64d84796c0291c720f8726
Reviewed-on: https://review.couchbase.org/c/couchstore/+/182040
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>