Support running python tests under TSan
Previously if building with TSan, the python tests would fail with:

    OSError: /opt/gcc-10.2.0/lib64/libtsan.so.0: cannot allocate memory in static TLS block

This is a somewhat cryptic error message, but is due to the TSan
runtime library being loaded to late in the process' startup - it must
be loaded early on.

For programs we comple ourselves this isn't a problem - they link to
libstsan.so and so it will be loaded when the process is first
started. However for python it does not link to tsan.so, and hence
tsan.so is only loaded when libcouchstore.so is later dlopen()'d via
the "import couchstore" python statement.

Fix by loading tsan.so on process startup via LD_PRELOAD for these
tests.

Note that the Python tests _also_ don't work under UBSan; this is a
related issue but is less straightforward to fix as the UBSan runtime
is currently linked statically and hence we can't simply preload it
like the TSan one. For now skip running these tests under UBSan.

Change-Id: I77c8500d042badf02930dacad17d10027864e92f
Reviewed-on: https://review.couchbase.org/c/couchstore/+/182047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>