MB-57298: [BP] Fix handling of couch_compact --dropdeletes
As part of refactoring to support concurrent couchstore compaction
(https://review.couchbase.org/c/couchstore/+/125688), the
--dropdeletes flag of couch_compact was inadvertently broken; due
to the semantics of the COUCHSTORE_COMPACT_FLAG_DROP_DELETES
compact_flag - quoting from couch_db.h:

    /**
     * A compactor hook will be given each DocInfo, and can either keep or drop
     * the item based on its contents.
...
     * If a compactor hook is set, COUCHSTORE_COMPACT_FLAG_DROP_DELETES will
     * *not* drop deletes, but will bump the purge counter. The hook is
     * responsible for dropping deletes.

As part of that refactor, we changed to use std::function with a
lambda wrapping the C-style filter API API, this in effect meant that
a compactor hook was _always_ set - the new implementation of the
C-style couchstore_compact_db_ex() function:

couchstore_error_t couchstore_compact_db_ex(Db* source,
                                                const char* target_filename,
                                                couchstore_compact_flags flags,
                                                couchstore_compact_hook hook,
                                                couchstore_docinfo_hook dhook,
                                                void* hook_ctx,
                                                FileOpsInterface* ops,
                                                PrecommitHook precommitHook) {
        return cb::couchstore::compact(
                *source,
                target_filename,
                flags,
                [hook, hook_ctx](Db& db, DocInfo* info, sized_buf body) -> int {
                    if (hook == nullptr) {
                        return COUCHSTORE_COMPACT_KEEP_ITEM;
                    }
                    return hook(&db, info, body, hook_ctx);
                },
    ...

i.e. note the 4th argument to compact() is a lambda funciton and hence
always non-empty, even if `hook` is null. Looking at the later
checking of the flag, note it simply checks that ctx->filterCallback -
the std::function - is non-empty:

    if ((bpWithDeleted & BP_DELETED_FLAG) && (!ctx->filterCallback) &&
        (ctx->flags & COUCHSTORE_COMPACT_FLAG_DROP_DELETES)) {
        return COUCHSTORE_SUCCESS;
    }

As such, this if() statement is never taken after the aforementioned
refactor.

This issue was flagged when upgrading to GCC-12 (CBD-5141), which as
part of the new container image, the lua interpreter was included,
which was not the case in the previous ubuntu-based image. As such,
the couchstore tests written in lua are now run in CV and this bug was
flagged by the dropdel test, which expects to see the --dropdeletes
flag function as expected:

    19/22 Test  #6: couchstore-dropdel ..................***Failed    4.89 sec
     Compacted /tmp/lua_N7i8gT -> /tmp/lua_XKR9gL
     Deletion dropping test: FAIL (/home/couchbase/jenkins/workspace/couchstore.linux_master/couchstore/tests/dropdel.lua:17: Expected 99999 results, found 100001)

Fix by removing the inline lambdas and only making the std::functions
non-empty if the C-style hook functions were also non-empty.

Change-Id: Ib14a350b330fd28c37b05a5d0dae4458bd56d809

---

Note that this has limited product impact - `couch_compact` is not
directly used in normal operation, compaction is driven internally
from ep-engine (via the C++ APIs), and moreover the
COUCHSTORE_COMPACT_FLAG_DROP_DELETES flag is _only_ used by
`couch_compact`. As such, this just means that support /
development-initiated compaction via the command-line tools would not
respect the --dropdeletes flag.

Change-Id: Id216247293eff060d78248ed61027b375e88a1cc
Reviewed-on: https://review.couchbase.org/c/couchstore/+/223477
Well-Formed: Restriction Checker
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>