MB-59379: [BP] Fix stack-use-after-scope
save_bulk had (an unused) functionality to set additional data
to the documents it would store. That unused functionality stored
the extra data in a variable stored on the stack which was out
of scope when the docinfo was passed to couchstore.

Given that the functionality isn't in use anywhere I removed
the erroronous code rather than trying to fix it (by allocating
more memory)

Change-Id: I98329d7aa1d298a21243addc66dff60d8c8b1a50
Reviewed-on: https://review.couchbase.org/c/couchstore/+/223481
Well-Formed: Restriction Checker
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>