MB-59041: Delete Incorrect Snappy Unit Tests
This commit deletes the unit tests:
1. `test_index_headers_v1`
2. `test_index_headers_v2`

These unit tests started to consistently fail after the upgrade to
snappy to version 1.10.0 (from 1.8.2) due to snappy producing different
results.

As per online references, the output of Snappy compression will not be
the same across library versions (and possibly not even across
architectures). The compressed output is an implementation detail and
not an interface. Hence, testing it is a bad idea as it can break at
some point (as it did after the recent Snappy upgrade in MB-57576).

These failing tests perform comparisons with the hardcoded results that
are compressed results produced by older versions of snappy.
Some of the hardcoded results date back to 2013.
Note that this test is run across different platforms with the same
expected compressed output hardcoded in test; this makes the test
incorrect.

We must only care about the data not getting corrupted in a
Snappy compress-decompress cycle.

Change-Id: Ie005c8fe1a4889034af50863304719da26aedf05
Reviewed-on: https://review.couchbase.org/c/couchstore/+/202487
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>