MB-61299: Bail out if fatbuf_get() returns null.
The memory location requested here is allocated in update_indexes
(couch_save.cc) & seems correct. But there is a possibility of a bug
elsewhere which might move the buf->pos beyond the initial requested
size & cause a buffer overflow (in which case fatbuf_get) returns NULL.

Bailout if that happens instead of crashing the memcached process.

Change-Id: I6111bcb28d9e05a01636577f140603537f789f84
Reviewed-on: https://review.couchbase.org/c/couchstore/+/209157
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>