MB-62188: Add encryption metadata when creating an encrypted file
To simplify key management, a per file random key is generated and
encrypted with a master key, with the master key identifier stored
in the metadata header.

The metadata header appears as the first chunk of the first block,
provided that it is a metadata block (type 0x02).

We bump the file version to 14 to signal that the file supports
encryption and may contain a metadata header.

Change-Id: Ie7959eefd2b085395c3e2ca9e3d9069d829ac4e0
Reviewed-on: https://review.couchbase.org/c/couchstore/+/211582
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>