MB-62188: Implement data chunk encryption
As data chunks are always appended to the end of the file, we use the
offset in the file as a unique value for the nonce. A chunk that appears
at the beginning of a block may be read from the block boundary or one
byte ahead, as the first byte of a block is the block type which will be
skipped over. To ensure that the same nonce is used for encryption and
decryption, we need to adjust the nonce such that it is the same in
those two cases. If the offset is at the block boundary, we increment
the nonce by one.

The encryption algorithm has built-in tampering detection, so no extra
checksum is computed.

Change-Id: I2eb8ff2256f320570ad0121f9d8a7ece14c2e740
Reviewed-on: https://review.couchbase.org/c/couchstore/+/211583
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>