MB-62188: Read metadata header only if indicated by file header
Add a 1-bit flag to the file header that will indicate whether the
metadata header needs to be read. This saves one read op when it is not
present.

We shrink the previous header position field by one bit so that no extra
space is used.

Also improve error logging and error code uniformity.

Change-Id: I620bfe62b7b694caa59160cde281fc11b8d24ed2
Reviewed-on: https://review.couchbase.org/c/couchstore/+/220671
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>