MB-65486: Commit by changing one byte
The presence of a file header (recognised by the Header block type)
signifies a commit point. If the header has only been partially written
it will be detected as corrupt, and opening the database will fail.

Instead, first write the file header with a Data block type. At that
point the header will not be found when opening the database. After
succefull fsync, overwrite the block type to be Header and fsync again.
Since only one byte is changed, we avoid issues due to partial writes.

Change-Id: Ic158270b796a4ee801b768e786ff839be89b4f18
Reviewed-on: https://review.couchbase.org/c/couchstore/+/224426
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>