MB-68169: Fix quota calculation for node splitting
We were previously subtracting from mr_quota, which broke when that was
made size_t, as quota can be exceeded to write one more item or if not
enough items are written. This was expected to make mr_quota negative,
which is not possible with size_t. As a result, node splitting would
usually fail, resulting in large nodes and degrading performance.

Fix is to use a separate variable for used quota and compare to the
target.

Change-Id: Idba6699a17c69c4cd2d47f18cbc1a2507aebbfe8
Reviewed-on: https://review.couchbase.org/c/couchstore/+/232706
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>