MB-69481: Use O_NOATIME when opening a file in read-only mode
Updating the file access time results in disk writes even when accessing
the file in read-only mode. On nodes with limited IOPS (e.g. cloud),
this can reduce performance. While users could mount their data volume
with the option to not update the access time, this is a more advanced
configuration that would affect all files in the volume. Couchstore
files are internal database files, so we can choose to not update the
access time as needed.

Change-Id: Ic0263e05cf8e444d2358c6af25587ea460e4f8da
Reviewed-on: https://review.couchbase.org/c/couchstore/+/236801
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>